Fixed bug 114768: GTK character mode doesn't support IBMgraphics.
I've added support for a map font to be specified in the GTK rc file as eg.,:

style "map"
{
	font = "vga"
}

widget "SlashEM map*" style "map"

If the font specified is encoded using the CP 437 character set then
IBMgraphics can be set and will work correctly. On RedHat Linux 6.1
the "vga" font is suitable for this use. I imagine there is at least
one such font under MS-Windows (note: CP 850 fonts will almost, but
not quite work. The fountain glyph, at least, is different).

I've also added code to detect if there are glyphs that the game will use
that are not present in the map font. A warning message is then issued.
Hopefully, this will make the problem clear if you try to run with IBMgraphics
set and with a non-CP 437 map font.