GtkHack: Profile versioning
This mainly affects UNIX profiles which are normally stored in ~/.gtkhackrc.
GtkHack will now look for a version-specific profile (~/.gtkhackrc-<version>)
before looking for the generic profile and use it if found. If the generic
profile is read but the version number contained in the profile is either
not a string or the string does not start with a major and minor version
number or if the version number is incompatible, then an error message is
issued and the profile read is aborted. GtkHack will then write out a
version-specific profile instead of overwriting the generic one.

This will allow multiple versions of gtkhack to coexist without interference.