Fixed bug 992244: Alt Gr key doesn't work
Fixed by reverting Warren's change of Apr 2002. It's not clear what this
was intended to achieve. The comment simply says "WIN32 left hand alt key".
Clearly it was intended to add support for using Alt as a meta key, but
this is already handled by the vanilla win32 interface in NetHack/Slash'EM
mode so it would appear to be superfluous.