Add some autogenerated doc files to fix windows builds
These were generated with "make distrib" in doc/, with the unix
build system; they're marked as ignored because they're not really
needed, and they can change slightly, depending on your configuration
and where you build them. We have to remember to regenerate them
if we end up changing the documentation; it could happen with
Guidebook.mn, for instance.