Providing players with a clue to the locations of alignment quest portals
They were way too damn hard to find, as they can randomly appear
anywhere from dlvl15 to dlvl39 and you'd have to search all of those
with absolutely no clue at all.
The message is the same for all three quests though, so you only know
that a quest entrance is there but not which one.