Eating certain things can influence the player's luck.
There's a chance of a +1 luck boost when eating the corpse of a
luck-draining monster now. Also, fortune cookies can modify the player's
luck based on their BUC; it is almost always a good idea to read the
fortune now.
Reading a strip of paper that appeared in some random food will also
modify the player's luck. Unless the food was cursed, it's always a good
idea to read that too.