SAVEBREAK 1.08: Added ability for players to play as hybrid races.
Probably full of bugs.

Known issue: The hybrid race selection prompts only appear if the "tty"
interface is used. Support for other interfaces might be added in future
versions.

There are currently 7 available hybridization races; the player can add
any number from 0 to 7. It's possible to do "self-hybridization", which
wasn't supposed to be there but I decided to declare it a feature.