SAVEBREAK 1.20: The big monst.c overhaul/cleanup (first part)
Figured that it's time to finally do this: adjust the spawn frequencies
for all the monsters so the game is a lot less unfair now. This isn't
even close to being finished though, but at least some progress have
been made. The immediately obvious result for the player is that special
rooms will have high-level monsters less often, and also  the general
dungeon population can hopefully be handled more easily now.