Untameable traitor monsters will betray the player more often.
This is being done to make sure that if a player somehow manages to tame
them (e.g. by polymorphing them into something tameable and waiting for
the polymorph to revert), they will be a lot less reliable.

Actually, this is an "obvious rule patch" to reign in the
overpoweredness of tame watchman-line monsters. Which would level up
quickly and get very powerful, plus the player is almost guaranteed to
encounter some early.