SAVEBREAK 1.25: Penalize hangup
Since hangup can be used for some major cheating, and I can't fix the
root issue due to the functions being complicated, using hangup is
penalized instead. All players beware: *DO NOT* use hangup if you want
to save! The penalty isn't instantly fatal, but it includes paralysis to
make sure that hangup cheating while surrounded by monsters is
definitely a bad idea. Of course this means that if the terminal hangs
up the player is penalized too; maybe in future versions there will be a
system that detects if hangup happened while a prompt was displayed. For
the time being though, players should just avoid hangup.