Unidentified items can merge onto (partly) identified stacks.
This will greatly facilitate the identifying game. Examples:

- Player has a known-cursed scroll of enchant armor. If they pick up
another one that is also cursed, but not known that it is cursed, it
will merge onto the first one generating a stack of 2 known-cursed
scrolls of enchant armor, thereby identifying the latter one's BUC
status.
- Player has 6 +0 daggers, which are known to be +0. If the player picks
up another stack of 4 daggers of which the enchantment isn't known, they
will merge onto the known-+0 stack if they're otherwise the same, so the
player won't have to individually identify the enchantment value of the
new stack.

Of course the general rule still applies, which is that only exactly
equal items can merge. And (partly) identified ones won't merge onto
non-identified ones because that would remove the player's knowledge
about them.