Players with the Amulet of Yendor can no longer cheat to get a spell off.
Now, they can always try if their mana is enough to pay for the normal
cost of the spell. The increased cost from having the Amulet is
calculated afterwards, and if it turns out that the spell then costs
more than the player actually has, it fails, draining the player's
current mana to 0 and losing a turn. Otherwise, the player could just
keep trying until the spell randomly costs very little!