One new attack type and 28 new damage types for monsters.
For now, these are all "bonus" attacks that no normal monster will uses;
the only way the player can encounter them is if a monster with random
attacks gets them. This will change in future versions, however. Also,
these are only coded for melee attacks now. The next thing that needs to
be done is to enable all of them for gaze, engulf and passive attacks,
too.