Level drain nerf
According to FIQ, the drain for gain fix is bullshit. While that
certainly isn't really the case (the fix is necessary), some balance
needs to be put in; monsters with wands of draining are way too powerful
otherwise. So, being drained might only take off a fraction of the
player's XP instead, and drain a level if the player only has little XP
gained on the current experience level.