Monsters will always resist their breaths.
Yeah I know, this is supposed to be defined in monst.c... but there are
so many monsters in this game, I really don't want to go through the
entirety of that file just to see if every monster with a breath attack
has the corresponding resistance. Also, shambling horror type monsters
that happen to be generated with breath attacks certainly don't get the
resistance automatically, so this should fix it.