Fixed dumpfile names work again
At least for me, the /tmp/ stuff caused error messages on Windows. Yes,
the "tmp" folder was there; maybe it's because it was preceded by a /
character? Anyway, Unnethackplus saves the dumplogs in the base folder
too and it works marvelously there, so this is how we're gonna do it
too.