Fix wait command and wizmode autocomplete
According to FIQ, not being interrupted while waiting if a monster comes
into view is a "feature". Well I say it's bullshit, especially since it
used to work before key rebinding hit us. There you see it, SLASH'EM has
a sleek interface while vanilla doesn't! So this is part of the "make
everything work like it used to" project, which aims to restore all
command functionality that was there pre-rebinding. Wizmode commands
didn't get autocompleted and I found out why: somehow they are defined
twice in the internal command array??????? So I created a "hack" in the
autocomplete function: if more than one command is possible, but all
those commands are the same, then it will autocomplete them anyway. It
annoyed me to no end that I could no longer input #poly and just have
that automatically be extended to #polyself (which also used to work
marvelously pre-rebinding).

This is probably not the last time I'll have to restore commands that no
longer work correctly.