Make polymorphs saner
This needed to be done for a long time, and does the following:
- Uncommon monsters are less likely to get chosen, whether it's the
player polymorphing into a random monster or a monster polymorphing into
another one.
- Monsters are less likely to polymorph into things that are much higher
level (more than 10 levels higher) than themselves. i.e. a goblin (level
1) is unlikely to become a level 20 monster, and even more unlikely to
become a level 30 monster, but a level 25 bodak would have the normal
chance of becoming one of them.