Shopkeeper services and other fixes
Shops can now run out of services, and identify is no longer guaranteed.
Because of a bug(?), SLASH'EM shopkeepers always offered both basic and
premier ID services; this is also fixed now. And if you try to use the
200 rocks exploit, the shopkeeper can call the kops - this is not a bug;
you are responsible for the # of things you pick up in a shop! :-P