Enable SINKS unconditionally
The game probably wouldn't even compile without it, and the same is true
for many other defines whose effects are just so deeply ingrained into
the game that I ultimately decided to remove the defines.

Note that I'm still defining them in config.h - this is done on purpose,
as a "fail safe" so to speak, so if any part of the code still somehow
depends on them then it will correctly be compiled.