Lit lightsabers and bones levels
Deadwater "sniped" me with a bones level that crashed whenever I tried
to move. He died with a lit lightsaber, and apparently hit the chance
for it to be removed yet the game didn't remove the timer associated
with it, so it was pointing to an undefined object. Then I got that
bones level and the game wanted to give an error message but the %s
string pointing to the item's name was also pointing at an invalid
address so it would always crash. This fix isn't the best, but it should
prevent the crash.