Savegame error fix
This is not guaranteed to remove it, because the version where it
occurred was not compiled with -f and so gdb didn't give exact
locations. All it gave was the functions in which the segfaults happen,
but not the exact line in the function that the game is choking over,
and somehow newsym() is erroring too, probably because the dungeon
regrowth function is being called during saving? This might potentially
be fixed by making sure the newsym() call only happens if the game is
not currently being saved.