Really fix the fucking savegame error
At least I hope it actually does this time, but probably not, seeing how the bug just keeps reappearing every time. I've started a local 1.9.3 character, finally, and thought I could enjoy the game without this bug at last. YASDed a few attempts, then got a game going, cleared a bunch of levels, tried to save, BAM, savegame error. Thankfully I was paranoid enough to attach the debugger - in fact I'll not play without it attached anymore - and it told me the exact line where the error was happening, because this is FINALLY a version compiled with the -f (or was it -g?) flag. Anyway, it's still very weird. Apparently when the game saves the individual dungeon levels, "uwep" and "uswapwep" and all those other things have already been undefined completely, and so even the safety check that makes sure they exist doesn't help anymore... so those lines may simply not be run while the game is attempting to save, it seems.

Let's see how long it takes until the goddamn bug reappears this time. *sigh*