Fix riders
This isn't an obvious rule patch just because I splatted on Astral one tile away from the altar today. Nope, my previous 7 ascensions have been made despite the existence of this bug, and only now it has actually managed to kill one of my chars. The Seven Deadly Sins were affected too. Now neither of them are getting excessive amounts of HP just from resurrecting; they keep the amount they get initialized with, which can be excessively high but if not then it doesn't suddenly become so for no reason. (Maybe I should have made it so that the bugged formula is still used if the player is using soviet mode, because SLASHTHEM is the variant that deliberately re-inserts bugs that I fixed. Only ones that *I* fixed, mind you, because I am the antichrist and anything I do is evil and needs to get the REVERT REVERT REVERT treatment. Yes, I'm referring to the "martial arts has a 1 in 4 chance to only deal one point of damage" and "monster spawn increase timers get re-initialized if you save and load the gam