Scrap theft attack item removal
This was potentially the cause for broken o_id chains and subsequent
savegame corruption. Yes, removing content is still not a part of the
SLEX design philosophy, but if something just flat out doesn't work then
there might not be another way. And I don't know how to make this work
in a way that does not generate hideous crash bugs, so it's removed for
now.