Timer error messages
Curse the existence of the o_id crash! That one is even worse than a
segfault because gdb does not allow us to track it down - when saving,
the savegame file will be corrupted, but you're not told immediately,
because it seems to save like normal and only when you come back will
you be greeted by a "cant find o_id" message... I can only hope that the
steal.c reversions fixed it.