AutoOpen - Remove superfluous "There's a door in the way".
This was printing after "The door resists!" with auto-open,
and was eventually revealed to also print in the rare case when
the door does not resist (in which case the door is no longer in
the way). If auto-open is disabled, you still get "That door is
closed", which is descriptive enough.
Also rearranged so the crazy-blue-walls thing is checked before
auto-open, since this results in you crashing hard into doors and
losing HP (no I have no idea what it means either), and this being
SLEX, opportunities for random unexplainable death should always
come before UI conveniences.