Annotate and overview
This is an absolute barebones version of that patch - the only thing it
does is that it lets you annotate levels and then see your own
annotations.

It does not automatically annotate levels for a couple of reasons: one
is that slex is too complex for that to really work, but more
importantly, there are nasty traps. Coding it so that it checks against
all kinds of nasty trap effects to see whether stuff should be added to
the annotations while under their effect would be hell.

Apparently it was also supposed to be possible to review the maps of
previously visited levels, which I consider to be even more problematic
in light of nasty traps, so we just don't have that. Sorry.

But hey, you can #annotate levels now and then use #overview to see what
you wrote!