Improve Rodney's towers
Replaced the zoo with a random special room, and also added a couple
extra special rooms which are likewise random. Because those levels were
so incredibly dull before. Known issue: if something like an inside room
is made in the actual Wizard's Tower, it may have undiggable walls
blocking the player from moving ahead. This will be fixed in a future
version by adding a terrain type that is a wall but will be exempt from
sp_lev.c's (or wherever that really happens) function that makes walls
undiggable.