Get enough candles for the candelabrum
Vlad's tower now has a bunch, and the spawn chances for the different
types of candles have also been overhauled so you will have a realistic
chance of finding at least a few in case wax and tallow are marked as
"will never be generated". After all, not being able to complete the
game because you don't find enough candles would really suck.