Fix bugged handling of obj_is_local error
This should fix the "segfault panic", "hallucination crash-on-save", "cant find o_id" and similar bugs which kept corrupting saves. Mind you, I did not find the ACTUAL underlying bug, but there were errors in the way I made the game try to handle the bug, and these errors were corrupting the savegame. I managed locally to successfully restore a game crash in segfaulter mode with these changes, which were previously considered completely impossible to recover. However, since the segfaulter mode is the one mode where these crashes are actually supposed to occur, and segfaulter as well as missingno races double as valuable tools for me since they allow me to test the game's behavior in the case of segfaults both with and without wizard mode, I intentionally made the game use the bugged code path if and only if the player uses the segfaulter race.