Better message for flickerer egotype
Players often did not get it that their attacks were doing ZERO damage,
and plugged away instead of, DUH, using an attack wand or something.
This is slex, one-trick ponies have no place here. If you aren't
versatile, maybe you should play a game that's so easy that everyone can
win it in their sleep, like vanilla NetHack. :P (Sorry if this sounds
harsh, but again, this is slex, players are expected to use a variety of
ways to kill enemies)