Pet rebellion change
Monsters whose base level is 50 or higher are potentially quite powerful
as pets, so I gave them a small chance of treachery. And the
jonadab-script-generated ones are potentially very unbalanced, so they
have a much higher chance of treachery.

Also, treachery should no longer be equal to "oh well, I'll just re-tame
it every time it happens" - there's now a chance of the monster becoming
frenzied and thus unable to be tamed.