More rumors
Some of them are mocking YASDs and ragequits of actual games that were
played on the server. If you recognize yourself, don't tell anyone that
it's you who is referred to in the message unless you want the whole
world to know :P
Also, fixed a few old rumors that used to be true but had to be changed
in their wording to ensure they're really true.