Fix hallucination segfault panic
At last I found the root of that problem! It is, in fact, a SLASH'EM error. The game just never checked for the hallucinated objects whether they had contents, and upon deleting them, the contents weren't deleted with them. So if the contents were e.g. eggs or corpses and thus had timers, those timers would persist and point to bugged objects leading to all sorts of crashes.