Fix the curses interface
Curse the developer of this thing. Whose idea was to make it so that if
the terminal is too small for the dialog window it calls a fucking
PANIC??? Also, popup_dialog may not be turned off, since if you do,
engraving will give an undiagnosable segfault, gdb just says it occurs
at 00000000 (great). And more importantly, it would be a way for the
player to segfault the game on purpose, allowing segfault scumming to
completely trash game balance (just segfault any bad results away, e.g.
when a wish for a +3 item only gives +0, crash the game on purpose to
get the wish back and try again). It is therefore of utmost importance
that no such bugs exist.