Try to balance ether items
The contamination was just proceeding way too quickly; rikersan
playtested and it became obvious to me that no one can seriously use
ether equipment right now. So I greatly reduced the rate of
contamination. The new values are not set in stone either; if it turns
out that ether equipment is still unusable because of the contamination,
they will be tweaked further.