Change hangup logic
Ditched the original hangup penalty, although it can be compiled back
into the game with the new HANGUPPENALTY define. The player still gets
paralyzed for a few turns to make it difficult to cheat monsters'
attacks away, and hanging up at an instadeath will of course still greet
you with a DYWYPI when you come back.

The game now tracks how many times hangup was triggered, and will show
it in the dumplog.