Add more monsters
Including a wafer-thin mint suggested by NCommander. He originally
wanted it to be a comestible but I decided to make it a monster.

Also thanks to Chris for making the kamerels and umachs. Like I did with
the Coure/Firre/etc. eladrin, I decided that "argenach" is unique
enough, they don't need the "rilmani" addition to their name, just like
"Coure" doesn't need the "Eladrin" addition. The kamerels, on the other
hand, do need it; a monster just called "goldshine" would seem weird, so
it's okay for them to be called "goldshine kamerel".