Make blue monsters have the correct shade of blue
Yes FIQ, I know, I'm a noob who's not capable of changing the blue color
in my cmder or putty settings. In reality I know how to do so of course,
but I don't expect the average player to be able to do the same - after
all, players get stumped by the fact that controls default to vi-keys on
the server too, and certainly not all of them know how to fix that
problem either! I've seen at least half a dozen players futilely
struggle around and eventually give up because of that, so I can't very
well expect them to be able to figure out how to make the goddamn
"floating eye blue" brighter either...