Prevent wall regrowth on undiggable squares
Levels with mazewalks that go through undiggable sections will leave the
dug-out tiles undiggable, which is a major problem. A more elegant
solution would be to make it so that those tiles are flagged diggable
when walls are created on them, but for now, let's prevent walls from
being created there since they'd have the potential to make the game
unwinnable.