Wierdo curses mode changes
I had to do them because the windows curses binary would segfault upon
character generation, but the real reason may have been something else.
For a reason completely unknown to me, curses mode now suddenly tries to
get the dungeon description from "nhshare" instead of "nhushare", yet it
doesn't create that file when compiled and the one created by compiling
with tty doesn't have that (of course). So I had to rename the share
file generated by the curses compilation and that made it work.