Make it harder to accidentally skip past prompts
For example, if you're eating a tin and there's a lot of battle spam
until you finally get "Eat it?", your spacebarring past the spam can
make you discard the tin even if you would have wanted to eat it. This
is bad design, so I fixed that. Also, hitting escape, or a key that is
interpreted as escape even if it should not, will no longer dismiss
wishing prompts without confirmation.