Fix message that busts terminals
The control codes in it were probably responsible for occasional fucked
up terminal sessions, and if SLEX does such things then the almighty K2
might get angry. And an angry K2 is the last thing we need. So this
bugged message is removed - YES I ACTUALLY REMOVED SOMETHING AND
THEREFORE BROKE MY OWN DESIGN PHILOSOPHY :P But to make up for it, a ton
of other, non-bugged messages have been added.