Try to make a CSPRNG
This will hopefully shut up complaints along the lines of "if you find
the seed you can get infinite wishes". The code does not work on windows
because /dev/urandom does not exist there and trying to use it anyway
would result in non-randomness, so it will be disabled unless you're
playing on a public server because it's mainly meant to make Junethack
harder to exploit anyway.