Pokedex buffer overrun "fix"
I'm not happy with this, but if I try to use \n in place of the
append_newline function I get very weird linebreaks on Windows. I guess
now the pokedex will look bad on larger terminals but what can you do?
Anything is better than having a "buffer overrun on demand" bug in
there.