Add an amulet field to whereis
This is slex, theoretically it's conceivable to run into a situation
where you can't easily determine in-game whether you have the real
Amulet of Yendor and we don't want players to be able to cheat, yet
Beholder on #hardfought apparently needs such a field. So I guess we add
one that's simply always zero. *in before complaints start being flung
at me*