Bones character change
Sadly, my previous attempt at changing the format of bones filenames was a complete failure, you'd constantly get "THIS IS A FATAL BUG" on the server. So we have to make do with having a maximum of 26 bones pools for the different dungeon branches, no matter how many branches there are. As a consequence I  tried to logically group them together so branches of similar difficulty will share bones now.