Do some more biiiiiiitching
Junethack has ended and some variants have managed to piss me off completely. But at least there can now be 11 months of SLEX, and 11 months for the variant developers to fix their shit. I'll make sure to bitch about things that are still not fixed in June 2019 and refer to the fact that I told them about it back in 2018 :P