Don't show crystal water items if flying
Thanks to Antichthon for both reporting this issue and insisting that I take care of it. Originally the suggestion was to make the tiles highlighted, but I decided to solve it in this way instead. Known issue: if you start flying by e.g. putting on an amulet, the item remains visible until the tile is redrawn.