Fetish heels and velcro sandals suck less
Making the player move more slowly is all fine and dandy, but why would cumbersome footwear affect them when attacking or doing other non-movement-related stuff? The slowing gown, on the other hand, is meant to make the player's actions more cumbersome in general, and therefore the effect of that one stays.