SAVEBREAK 2.55: materials overhaul
This commit does something very useful: it changes the internal names of materials to avoid clashes, so e.g. IRON becomes MT_IRON. It took an hour and a half to replace all occurrences of materials strewn out over the code, but I'd say it's worth it :) Also added 7 new material types, but they don't actually do much yet and need to be finished later, as well as new items need to be added that use those materials.