FUCK THE SAVEGAME ERROR
It appears that if during dgn_growths the game tries to access something like "uimplant", there's also a chance of getting a segfault. gdb said that while uimplant did point to some memory address, "p uimplant->otyp" said that the memory could not be addressed, so that's probably why it crashed.