FREEPLAY AFTER ASCENSION :D
Yes, it's true! If you ascend now, you get the choice to keep playing your game. The ascension will appear in the xlogfile and high-score list, but you can then go on playing if you want and explore everything you've skipped! Going back to the Elemental Planes requires the player to go to Moloch's sanctum first, and then the planes are open forever; leaving them can then be done by simply pressing < on any high altar.

This is unfinished, I probably need to make it so that the Amulet of Yendor can no longer be used to ascend again, especially since monsters can ascend with it in evilvariant mode and end your game!