Fix bug where sometimes a monster or dungeon feature would be gray instead of whatever colour it had to be This occurred when it was to the right of an engraving, object pile, egotype monster, or something else that had a highlighted background.
This was a fun one!  Essentially, this is because nethack is overengineered (and you thought that was a modern phenomenon).  In tty, there's the tty layer and the termcap layer, each of which keeps track of the 'current colour' separately.
The tty layer tries to be smart, so if you have a brown object next to another brown object, you can say term_start_color(brown); draw(first object); draw(second object); term_end_color().  The way it does this is by saying 'if the current object is the same colour as the last one: draw it; otherwise: term_end_color(); term_start_color(new color)'.
The problem is that 'same colour as the last one' relies on the state in tty, while with 'draw(second object)', the colour of that object depends on the state in termcap
In this case, the code for highlighting looked like:   if(hilite) term_start_bgcolor(whatever); draw(object); if(hilite) term_end_color()
term_end_color() modifies termcap's state so it thinks that things drawn in the future should be drawn with no colour
But it doesn't modify tty's state
So on the next iteration, tty looks at its internal state and thinks "yeah, we've totally got a colour we can reuse"
but it's wrong; termcap thinks no such thing
Fixed by telling tty that termcap's colouring was reset, so it can't reuse the colouring from last time