Nerf Rodney's intervention
Again, longer dungeon length makes this necessary. Since there are so many levels in the ascension run now, including a lesser form of intervention when you're doing the Yendorian Tower, we need to reduce the amount of "fuck you" that you get. Also, the Zrutinator is very strong so I decided to make him appear only sometimes instead of always.