Do some work on spells
Petrify spell only works on adjacent targets now, and I reworked the calculations for spell failure rates but not sure whether I'm happy with them now. I'd do some testing and more work but a RL thunderstorm is blowing into my area so I guess it's time to call it a night.