Failsafes for mkobj routines
There's apparently a new bug where otyp NUM_OBJECTS can be requested to be generated in mksobj(), but I don't know where that call is coming from. This was also segfaulting on occasion because it returned a zero object even if the caller didn't check whether an object was returned or not; that should now be fixed. In any case, it will now try to create the usual gold piece, hopefully fixing the crash issues.