New option "simpledescs"
This shortens many words that are appended to items, such as "uncursed", "(being worn)" or "finalized". Default off. Can be turned on to make it more likely that the item names fit on one line in the inventory screen, because someone made it so that they don't wrap around to the next line (gah).