Hopefully better fix for demon gating
Players in their natural form can have M2_DEMON, but only certain races are supposed to be able to do demon gating. Instead of trying to blacklist all the races that could possibly be affected and then probably forgetting one anyway, let's just try to work with a whitelist instead.