Changes to erdrick and DQ slime quests
These are by Kahran042, although I had to make some adjustments because the des files weren't formatted properly, causing the compile to fail :P The old Erdrick quest maps will probably be used for sone new, to-be-created role, which saves me the trouble of having to design new maps for that. Thanks Kahran042!