Bad equipment shouldn't be ultra-deadly
The amulet of strangulation is so deadly, I had to disable it for that specific effect. Do not mistake this for an "Amy is getting soft" precedent - all other items, including the amulet of premature death, are still eligible. Strangulation in particular is just boring if you get it force-equipped on you because then it's "you either have one of the few methods available for removing it immediately or the game is over", which is NO fun.