Okay, that was the last straw
Due to the constnat proliferation of undiagnosable and unrecoverable crashes out of the blue, I hereby decide that the curses interface is disabled until someone can come up with a fix, which may well be never because I can't fix it myself. Then I can focus once again on fixing actual crashes, i.e. the ones that also happen in tty mode :P