Change youpoly for lycanthropes
It's just insane if you can switch back and forth whenever you want to, and doubly so that the mana cost for the potentially very powerful ak Thief is dead! is so laughably low. So the ability is on a timer now and the mana costs should also be better. On the plus side, using the ability trains the squeaking skill now.