New artifacts
"omg two out of the 2300 total artifacts are slightly offensive in the eyes of whiny sjws and therefore we invalidate the entire game" bullshit. I could never be such a depraved person, and I very certainly don't want to join their community again because slex is a banned topic there anyway. Either way, now we still need to add new monsters that use the base items.