More new monsters
Yeah I know, they're sooooo offensive and I'm such a bad, stupid, evil or whatever person. Blah to that I say, like I would care at this point. SLEX ain't gonna make a comeback on that server, that's for sure. Anyway, this is the last commit before my vacation ends, so from this point on I'll probably not be able to pump out nearly as many commits for a while.