Add even more artifacts
Insert mocking comments about SJWs here who get too hung up on some minor little detail and decide to dismiss all the hard work that was put into this huge game because of that. If that pettiness ain't SJW, I don't know what is. A normal person would just play a game or ignore it, not complain constantly just because the developer's moral concepts are different from the SJW's own ones.