The SLEXtinctionist
Their fault for coming up with the idea, it was just TOO good not to use. Also, when testing, it turned out that this monster was using its taunts way less than it was supposed to, and the ranged data delete attack almost never procced, so I wasted a lot of time investigating and found out that ranged_attk() returned FALSE if the monster doesn't have specific attacks! This is fixed now, and may have the side effect that many AT_HUGS monsters attack much more often than before, which works as intended because they were never supposed to be that reluctant to attack the player.