Monsters
Making it so that I can't star the repo anymore is a dick move, by the way. What are you afraid of? Me possibly defiling the repo by writing a comment or something? Because I wasn't planning to do so anyway, and I can still access the repo anyway, it's just more annoying because I now have to use an additional link to do so.