New monsters
Many of these are stationary, for two reasons: one, it should reduce the proliferation of monsters who chase the player aggressively, easing on some of the stress commonly encountered in wide open spaces and making it less likely for the player to become surrounded; two, they're usable for symbiosis, which is supposed to become more attractive in future versions. I'm planning to add more such monsters to increase the variety of potential symbiotes.