Fix STUPID relocation
Gah! For regular teleports, which by definition can take place on teleport-enabled levels only, it's obviously okay if the player can end up in a wall as a phasing monster, because it's easy to just teleport back. But if you get relocated on a sokoban level clone in the forging chamber, you can't simply teleport back out...