Super Releasio Nethack :-P
This race isn't really meant to be playable. The releasier has permanent autopilot and mojibake, glyphs have a chance of being recolored red, cyan or white for them, and they start with a randomized skill set, but the amount and caps of starting skills are taken into account.

No one will play this race, ever. Even if the player base was as big as it used to be, no one would play it. And those who accidentally pick it or received it by going random will play for 50 turns tops, then quit.