Heavily nerf Elbereth and similar cheese
Elbereth, scare monster and temple sanctuary are capable of rendering the game much easier than I want it to be, since they allow players to still freely attack the monsters while they're often forced to flee or otherwise incapable of harming the hero. That's very unfair in a way that favors the player, and therefore the chance for monsters to be affected will now decrease every time they are affected. As a result, using Elbereth excessively will make it less likely to work until it reaches a point where it's just plain not worth it anymore.