Messages
This is what I have to say to people who really think I'd be grateful if they make deconstructed variants of my game (already a sacrilege) and then expect me to be fucking grateful (inexcusable insult). Anyway, the person in question has been added to the list(TM) now, and no apologies or excuses will be considered.

For me, this annoyance has been dealt with now, any attempt to mention such variants or their creators again will result in immediate and permanent bans from the IRC and subreddit for the perpetrators. And now that this is done, we'll once again focus on the only game that matters, which is SLEX, of course.