Don't warp pets in freeplay mode
No, this was not a "reward", in fact it was rather detrimental because I would not be able to purposefully park pets on other items so they don't kill the nurses or other useful monsters on the level I was going to. Also there's not supposed to be rewards for freeplay mode; the mode is strictly there so that you can keep playing after you've ascended, it is not meant as a power trip mode where your char is suddenly all-powerful. (And no, it's also not meant to be a slightly-more-powerful mode; currently, the only gameplay change in freeplay mode is that the riders and sins no longer leave corpses, and that's because we both want to prevent endless farming and also prevent them from endlessly getting on the player's nerves.)