Nethackwiki is dead
They haven't completed the purge yet, but that's probably just a matter of time. Fucking mob of bigoted haters who don't tolerate anything that doesn't match their standards. They're going to erase SLEX from history, I know it. Every resource on the internet that lists NetHack variants will just not mention that SLEX is a thing. That feels like something the Chinese government would do with things they consider to be in violation of the Communist Party's stance.