Pompeji squads
These will thwart ziratha's strategy of murdering all the shopkeepers and temple priests, as doing so will now cause the player character to be chased by the pompeji squad for a long time. Pompejis are like superpowered keystone kops with above-average stats for their level, high damage output and good resistances, some of them also reflect, and they should be sent with the intention of killing those who disrput the Yendorian infrastructure! :-P