Code most of the new traps
Bah, this is such a fucking pain in the butt, and I didn't even finish yet (the monster generator remains to be done)! But anyway all the cunts who hate my game because of a few rare "offensive" features can all jump off a bridge because they don't even know how much work it is to make a NetHack variant. If they were to construct a castle in real life, I'll say it sucks just to spite the amount of work they put into it.