Reticulan downside
Sorry. But let's list the downsides:
- certain relatively annoying monster types are slightly more common
- that's all. Yep, only one downside. Except if you count "starts in the space base" as one.
And the benefits:
- lotsa Pw growth per level
- several rare intrinsics
- buncha good techniques
- doesn't have to fear cannibalism or same-race sacrifice
- higher than average starting stat roll
- and this is the kicker, a fucking INERTIA melee attack that few things can resist.

So yeah. Aggravate monster and the occasional evilroom spawn should even the playing field a little. All this really does is to make the reticulan encounter some more monsters, so it even lets them kill more stuff for XP and loot, but at least some of these monsters will hopefully be not a complete pushover.