Fix the hangup handler
Thanks amateurhour for fixing it. This has the side effect that SLASH'EM and several other variants on the enemy server also received the fix, which I had to accept in this case because I really wanted the bug gone from slex; if I had managed to figure out this fix on my own, I'd not have told the people from that other server about it, they'd have had to get it themselves. Reason: they are all under the order of their admin, and the order reads "do not talk to Amy, in fact do not even acknowledge her presence". Bastards. I won't die from that treatment, you hear?