Change stat training
Exercise is now less likely to increase stats which are already quite high (and of course cannot increase them beyond 18), but there's alternate ways to increase stats being added, and one of the reasons for this is that CHA and INT shouldn't be so ultra difficult to increase. There need to be ways to get those up to usable values, too!