Fix brain-dead programming
This is an ugly hack: because they couldn't be arsed to simply make a SPFX_DBLDAMAGE property and it would be way too much work to add one now (since all artifacts that also have a to-hit bonus work properly and it's just the few without one that don't), I decided to solve things this way.

This has been tested to ensure it doesn't break saves, but unfortunately as the artifacts array is created and saved into the savegame file upon starting the game, existing characters won't benefit from this change and will have the old, bugged versions of those artifacts.