Redo the missingno race
Their polymorphing was originally meant to result in wacky randomized monsters but that broke when we changed the method for creating the monsters array, so it was mostly non-functional since then. This commit makes them functional again, although in a way that's quite different from the original. The item and trap associated with the missingno also work again now.