Finish the motherfucking artifacts
Gaaaaaaaaaaaaah. I don't want to have to see any nethack code for the next month now because this was such a pain in the butt. Also, I want to sleep for 25 hours straight and have a bonus of €200 transferred to my account as a compensation for all the hard work I put into coding those artifacts.