Close an exploitable loophole
Thanks Porkman for showing me back in the day that unless you fireproof your game against such exploits, there WILL be players determined to abuse the hell out of them and completely trample the carefully constructed balance of your game. You showed me that I'm not paranoid after all and that the players are really trying to exploit the hell out of my game just so they can score an easy win. So my act of closing such loopholes on sight isn't a mental illness that warrants me to visit a doctor, but completely natural behavior. :-)