Fix a SLASH'EM bug with kicking
This bug has existed in SLASH'EM ever since they added the "need +X weapon to hit" attribute: if your boots have an enchantment value of -1 or worse, they cannot hurt any monster because all monsters count as "needs +0 or higher enchantment to hit" for that purpose due to this oversight. It doesn't do the same for weapons, a -1 weapon can still hit things, so... it's fixed now.